+++ test/src/Repository/CraftingCategoryRepositoryTest.php
@@ -120,8 +120,7 @@ class CraftingCategoryRepositoryTest extends TestCase
                    ->method('removeIds')
                    ->with($orphanedIds);
 
-        $result = $repository->removeOrphans();
-        $this->assertSame($repository, $result);
+        $repository->removeOrphans();
     }
 
     /**
@@ -239,7 +238,6 @@ class CraftingCategoryRepositoryTest extends TestCase
                    ->method('getEntityName')
                    ->willReturn($entityName);
 
-        $result = $this->invokeMethod($repository, 'removeIds', $craftingCategoryIds);
-        $this->assertSame($repository, $result);
+        $this->invokeMethod($repository, 'removeIds', $craftingCategoryIds);
     }
 }