+++ test/src/Repository/MachineRepositoryTest.php
@@ -331,8 +331,7 @@ class MachineRepositoryTest extends TestCase
                    ->method('removeIds')
                    ->with($orphanedIds);
 
-        $result = $repository->removeOrphans();
-        $this->assertSame($repository, $result);
+        $repository->removeOrphans();
     }
 
     /**
@@ -444,7 +443,6 @@ class MachineRepositoryTest extends TestCase
                    ->method('getEntityName')
                    ->willReturn($entityName);
 
-        $result = $this->invokeMethod($repository, 'removeIds', $machineIds);
-        $this->assertSame($repository, $result);
+        $this->invokeMethod($repository, 'removeIds', $machineIds);
     }
 }